console: fix warning on newer gcc versions
git-svn-id: https://svn.code.sf.net/p/moonunit/svn/trunk/moonunit@323 7cfde7b9-103a-0410-917d-a4b0b378e324

+++ src/plugins/console/console.c
@@ -551,7 +551,7 @@ set_details(ConsoleLogger* self, bool details)
 static const char*
 get_loglevel(ConsoleLogger* self)
 {
-    switch (self->loglevel)
+    switch ((int) self->loglevel)
     {
     case -1:
         return "none";
